requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.10.0
black==22.1.0
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
flake8==4.0.1
Flask==2.0.3
frozenlist==1.3.0
future==0.18.2
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
line-bot-sdk==2.1.0
MarkupSafe==2.1.1
mccabe==0.6.1
multidict==6.0.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
numpy==1.22.2
oauthlib==3.2.0
pandas==1.4.0
pathspec==0.9.0
platformdirs==2.4.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyFirmata==1.1.0
pyflakes==2.4.0
pyparsing==3.0.7
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
six==1.16.0
soupsieve==2.3.1
tableauserverclient==0.17.0
toml==0.10.2
tomli==2.0.0
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
yarl==1.7.2
